cart_broker/requirements.txt
bleak==2.1.1
dataclasses==0.6
dbus-fast==3.1.2
dotenv==0.9.9
paho-mqtt==2.1.0
python-dotenv==1.2.1